requirements.txt
streamlit==1.44.1
openai>=1.51.0
rich>=13.9.0
nvidia-nat>=1.5.0
